requirements.txt
streamlit
supabase
PyPDF2
requests
python-dotenv
chromadb
sentence-transformers
openai